+++ src/main/java/net/chaos/chaosmod/gui/tweaker/InitGuiEventHandler.java
@@ -28,8 +28,6 @@ public class InitGuiEventHandler {
 //		}
 
 		if (event.getGui() instanceof GuiMainMenu) {
-//			Minecraft.getMinecraft().getSoundHandler().stopSounds();
-//			playSound(ModSounds.BLACK_ROVER, 1.0f);
 		}
 	}
 
@@ -57,8 +55,6 @@ public class InitGuiEventHandler {
 	// FIXME : playlist issue
 	@SubscribeEvent
 	public static void onSoundPlayed(PlaySoundEvent event) {
-//		System.out.println("played sound " + event.getName());
-//		System.out.println("tried to check current sound " + ClientSoundHandler.getSoundName());
 		if (event.getName().equalsIgnoreCase(ClientSoundHandler.getSoundName())) {
 	        if (ClientSoundHandler.isMusicPaused()) {
 	            ClientSoundHandler.forcePauseMusic();